refactor: Component 디랙토리 정리
각 컴포넌트는 해당 story와 같은 레벨에 존재하도록 한다.
