+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import Button from "components/common/Button";
+import Button from "components/common/Button/Button";
 import Input from "components/common/Input";
-import Image from "components/common/Image";
+import Image from "components/common/Image/Image";
 
 export default function App() {
   return (
