+++ src/components/common/Button/Button.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, StoryObj } from "@storybook/react/*";
-import Button from "../Button";
+import Button from "./Button";
 
 export default {
   title: "Components/Button",
