+++ src/components/common/Form/Input/Input.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta } from "@storybook/react/*";
-import Input from "../Input";
+import Input from "./Input";
 
 export default {
   title: "Components/Input",
