+++ src/components/common/GoodItem/Default/Default.stories.tsx
