+++ src/components/common/GoodItem/Ranging/Ranging.tsx
